requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.34.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
